finish identity server

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -40,12 +41,13 @@ public class AuctionsController(AuctionDbContext context, IMapper mapper, IPubli
         return mapper.Map<AuctionDto>(auction);
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
     {
         var auction = mapper.Map<Auction>(auctionDto);
-        // TODO: add current user as seller
-        auction.Seller = "test";
+        
+        auction.Seller = User.Identity?.Name ?? throw new InvalidOperationException("User not found");
         
         context.Auctions.Add(auction);
         
@@ -60,6 +62,7 @@ public class AuctionsController(AuctionDbContext context, IMapper mapper, IPubli
         return CreatedAtAction(nameof(GetAuctionById), new { id = auction.Id }, mapper.Map<AuctionDto>(auction));
     }
 
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult<AuctionDto>> UpdateAuction(Guid id, [FromBody] UpdateAuctionDto auctionDto)
     {
@@ -71,7 +74,8 @@ public class AuctionsController(AuctionDbContext context, IMapper mapper, IPubli
             return NotFound();
         }
 
-        // TODO: check seller == username
+        if (auction.Seller != User.Identity?.Name) return Forbid();
+        
         auction.Item.Make = auctionDto.Make ?? auction.Item.Make;
         auction.Item.Model = auctionDto.Model ?? auction.Item.Model;
         auction.Item.Year = auctionDto.Year;
@@ -90,6 +94,7 @@ public class AuctionsController(AuctionDbContext context, IMapper mapper, IPubli
         return Ok();
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
@@ -99,7 +104,7 @@ public class AuctionsController(AuctionDbContext context, IMapper mapper, IPubli
             return NotFound();
         }
 
-        // if (auction.Seller != User.Identity?.Name) return Forbid();
+        if (auction.Seller != User.Identity?.Name) return Forbid();
         context.Auctions.Remove(auction);
         
         await publishEndpoint.Publish<AuctionDeleted>(new {Id = auction.Id.ToString()});
